Extracted Table Cells from multiple table components into a singular TableCellComponents File where each table imports as required
