+++ pokedex-react-flask-app/src/components/PokemonZMoveTable/PokemonZMoveTable.jsx
@@ -1,76 +1,20 @@
 import React from "react";
-import { useNavigate } from "react-router-dom";
 
-import { formatName } from "../../helpers/format";
-import { getItemSprite } from "../../helpers/pictures";
-import { handleItemError } from "../../helpers/error";
 import { modifyPokemonUniqueZMove } from "../../helpers/modifyForTable";
 
 import { Table } from "../";
+import {
+  MoveNameComponent,
+  MoveComponent,
+  TypeImageComponent,
+  KindImageComponent,
+  ItemComponent,
+} from "../TableCellComponents/TableCellComponents";
 
 const PokemonZMoveTable = ({ move }) => {
-  const navigate = useNavigate();
-
-  const NameComponent = ({ value, row }) => {
-    return (
-      <p
-        className="move-list-item-name clickable"
-        onClick={() => navigate(`/moves/${row.original.id}`)}
-      >
-        {formatName(value)}
-      </p>
-    );
-  };
-  const MoveComponent = ({ value }) => {
-    return (
-      <p className="clickable" onClick={() => navigate(`/moves/${value.id}`)}>
-        {formatName(value.name)}
-      </p>
-    );
-  };
-  const TypeImageComponent = ({ value }) => {
-    return (
-      <img
-        className="table-image clickable"
-        src={`/icons/types/${value.name}.png`}
-        alt={`${value} icon`}
-        onClick={() => navigate(`/types/${value.id}`)}
-      />
-    );
-  };
-
-  const KindImageComponent = ({ value }) => {
-    return (
-      <img
-        className="table-image"
-        src={`/icons/kinds/${value}.png`}
-        alt={`${value} icon`}
-      />
-    );
-  };
-  const ItemComponent = ({ value }) => {
-    return (
-      <div
-        className="clickable"
-        style={{
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-        onClick={() => navigate(`/items/${value.id}`)}
-      >
-        <p style={{ marginRight: "0.5rem" }}>{formatName(value.name)}</p>
-        <img
-          src={getItemSprite(`${value.name}--bag`)}
-          onError={handleItemError}
-        />
-      </div>
-    );
-  };
-
   const { tableData, columns } = modifyPokemonUniqueZMove({
     move,
-    NameComponent,
+    NameComponent: MoveNameComponent,
     MoveComponent,
     TypeImageComponent,
     KindImageComponent,
@@ -79,7 +23,7 @@ const PokemonZMoveTable = ({ move }) => {
 
   return (
     <div>
-        <h4 className="text-center">Z-Moves</h4>
+      <h4 className="text-center">Z-Moves</h4>
       <Table
         data={tableData}
         columns={columns}
