+++ pokedex-react-flask-app/src/components/Types/Types.jsx
@@ -3,18 +3,16 @@ import { useNavigate } from "react-router-dom";
 
 import "./Types.scss";
 
-const Types = ({ types, pageTypeId = null }) => {
+const Types = ({ types, pageId = null }) => {
   const navigate = useNavigate();
   return (
     <ul className="type-list">
       {types.map((type) => {
         return (
           <li
-            className={
-              pageTypeId !== type.pokemon_v2_type.id ? "clickable" : ""
-            }
+            className={pageId !== type.pokemon_v2_type.id ? "clickable" : ""}
             onClick={() => {
-              if (pageTypeId !== type.pokemon_v2_type.id) {
+              if (pageId !== type.pokemon_v2_type.id) {
                 navigate(`/types/${type.pokemon_v2_type.id}`);
               }
             }}
